Test in test-users.js to reject untrimmed passwords was failing.  I found the root of the problem to be in users/router.js line 42.  Originially, the variable "nonTrimmedField" was set equalto "explicitlyTrimmedFields.find( >     field => req.body[field].trim() !== req.body );
I found that the problem was not specifically checking for inequality to req.body[field] which was causing the first item in the request body (username) to always come back as the location of the error, even if it was in fact trimmed.  The variable is now set as follows:

  const nonTrimmedField = explicitlyTrimmedFields.find(
    field => req.body[field].trim() !== req.body[field]
  );

Which has enabled the test to pass.